refactor: split HomePage
