+++ frontend/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useHTMLElement';
 export * from './useForkRef';
+export * from './useIsMobileView';
